+++ lib/ets/set.ex
@@ -78,10 +78,8 @@ defmodule ETS.Set do
   """
   use ETS.Utils
 
-  alias ETS.{
-    Base,
-    Set
-  }
+  alias ETS.Base
+  alias ETS.Set
 
   @type t :: %__MODULE__{
           info: keyword(),
@@ -409,6 +407,110 @@ defmodule ETS.Set do
   """
   @spec match!(any()) :: {[tuple()], any() | :end_of_table}
   def match!(continuation), do: unwrap_or_raise(match(continuation))
+
+  @doc """
+  Deletes all records that match the specified pattern.
+
+  Always returns `:ok`, regardless of whether anything was deleted or not.
+
+  ## Examples
+
+      iex> set = Set.new!(ordered: true)
+      iex> Set.put!(set, [{:a, :b, :c, :d}, {:e, :b, :f, :g}, {:h, :i, :j, :k}])
+      iex> Set.match_delete(set, {:_, :b, :_, :_})
+      {:ok, set}
+      iex> Set.to_list!(set)
+      [{:h, :i, :j, :k}]
+
+  """
+  @spec match_delete(Set.t(), ETS.match_pattern()) :: {:ok, Set.t()} | {:error, any()}
+  def match_delete(%Set{table: table} = set, pattern)
+      when is_atom(pattern) or is_tuple(pattern) do
+    with :ok <- Base.match_delete(table, pattern) do
+      {:ok, set}
+    end
+  end
+
+  @doc """
+  Same as `match_delete/2` but unwraps or raises on error.
+  """
+  @spec match_delete!(Set.t(), ETS.match_pattern()) :: Set.t()
+  def match_delete!(%Set{} = set, pattern) when is_atom(pattern) or is_tuple(pattern),
+    do: unwrap_or_raise(match_delete(set, pattern))
+
+  @doc """
+  Returns records in the specified Set that match the specified pattern.
+
+  For more information on the match pattern, see the [erlang documentation](http://erlang.org/doc/man/ets.html#match-2)
+
+  ## Examples
+
+      iex> Set.new!(ordered: true)
+      iex> |> Set.put!([{:a, :b, :c, :d}, {:e, :c, :f, :g}, {:h, :b, :i, :j}])
+      iex> |> Set.match_object({:"$1", :b, :"$2", :_})
+      {:ok, [{:a, :b, :c, :d}, {:h, :b, :i, :j}]}
+
+  """
+  @spec match_object(Set.t(), ETS.match_pattern()) :: {:ok, [tuple()]} | {:error, any()}
+  def match_object(%Set{table: table}, pattern) when is_atom(pattern) or is_tuple(pattern),
+    do: Base.match_object(table, pattern)
+
+  @doc """
+  Same as `match_object/2` but unwraps or raises on error.
+  """
+  @spec match_object!(Set.t(), ETS.match_pattern()) :: [tuple()]
+  def match_object!(%Set{} = set, pattern) when is_atom(pattern) or is_tuple(pattern),
+    do: unwrap_or_raise(match_object(set, pattern))
+
+  @doc """
+  Same as `match_object/2` but limits number of results to the specified limit.
+
+  ## Examples
+
+      iex> set = Set.new!(ordered: true)
+      iex> Set.put!(set, [{:a, :b, :c, :d}, {:e, :b, :f, :g}, {:h, :b, :i, :j}])
+      iex> {:ok, {results, _continuation}} = Set.match_object(set, {:"$1", :b, :"$2", :_}, 2)
+      iex> results
+      [{:a, :b, :c, :d}, {:e, :b, :f, :g}]
+
+  """
+  @spec match_object(Set.t(), ETS.match_pattern(), non_neg_integer()) ::
+          {:ok, {[tuple()], any() | :end_of_table}} | {:error, any()}
+  def match_object(%Set{table: table}, pattern, limit),
+    do: Base.match_object(table, pattern, limit)
+
+  @doc """
+  Same as `match_object/3` but unwraps or raises on error.
+  """
+  @spec match_object!(Set.t(), ETS.match_pattern(), non_neg_integer()) ::
+          {[tuple()], any() | :end_of_table}
+  def match_object!(%Set{} = set, pattern, limit),
+    do: unwrap_or_raise(match_object(set, pattern, limit))
+
+  @doc """
+  Matches next records from a match_object/3 or match_object/1 continuation.
+
+  ## Examples
+
+      iex> set = Set.new!(ordered: true)
+      iex> Set.put!(set, [{:a, :b, :c}, {:d, :b, :e}, {:f, :b, :g}, {:h, :b, :i}])
+      iex> {:ok, {results, continuation}} = Set.match_object(set, {:"$1", :b, :_}, 2)
+      iex> results
+      [{:a, :b, :c}, {:d, :b, :e}]
+      iex> {:ok, {results2, continuation2}} = Set.match_object(continuation)
+      iex> results2
+      [{:f, :b, :g}, {:h, :b, :i}]
+      iex> {:ok, {[], :end_of_table}} = Set.match_object(continuation2)
+
+  """
+  @spec match_object(any()) :: {:ok, {[tuple()], any() | :end_of_table}} | {:error, any()}
+  def match_object(continuation), do: Base.match_object(continuation)
+
+  @doc """
+  Same as `match_object/1` but unwraps or raises on error.
+  """
+  @spec match_object!(any()) :: {[tuple()], any() | :end_of_table}
+  def match_object!(continuation), do: unwrap_or_raise(match_object(continuation))
 
   @spec select(ETS.continuation()) ::
           {:ok, {[tuple()], ETS.continuation()} | ETS.end_of_table()} | {:error, any()}
@@ -794,4 +896,86 @@ defmodule ETS.Set do
   """
   @spec wrap_existing!(ETS.table_identifier()) :: Set.t()
   def wrap_existing!(table_identifier), do: unwrap_or_raise(wrap_existing(table_identifier))
+
+  @doc """
+  Transfers ownership of a Set to another process.
+
+  ## Examples
+
+      iex> set = Set.new!()
+      iex> receiver_pid = spawn(fn -> Set.accept() end)
+      iex> Set.give_away(set, receiver_pid)
+      {:ok, set}
+
+      iex> set = Set.new!()
+      iex> dead_pid = ETS.TestUtils.dead_pid()
+      iex> Set.give_away(set, dead_pid)
+      {:error, :recipient_not_alive}
+
+  """
+  @spec give_away(Set.t(), pid(), any()) :: {:ok, Set.t()} | {:error, any()}
+  def give_away(%Set{table: table} = set, pid, gift \\ []),
+    do: Base.give_away(table, pid, gift, set)
+
+  @doc """
+  Same as `give_away/3` but unwraps or raises on error.
+  """
+  @spec give_away!(Set.t(), pid(), any()) :: Set.t()
+  def give_away!(%Set{} = set, pid, gift \\ []),
+    do: unwrap_or_raise(give_away(set, pid, gift))
+
+  @doc """
+  Waits to accept ownership of a table after it is given away.  Successful receipt will
+  return `{:ok, %{set: set, from: from, gift: gift}}` where `from` is the pid of the previous
+  owner, and `gift` is any additional metadata sent with the table.
+
+  A timeout may be given in milliseconds, which will return `{:error, :timeout}` if reached.
+
+  See `give_away/3` for more information.
+  """
+  @spec accept() :: {:ok, Set.t(), pid(), any()} | {:error, any()}
+  def accept(timeout \\ :infinity) do
+    with {:ok, table, from, gift} <- Base.accept(timeout),
+         {:ok, set} <- Set.wrap_existing(table) do
+      {:ok, %{set: set, from: from, gift: gift}}
+    end
+  end
+
+  @doc """
+  For processes which may receive ownership of a Set unexpectedly - either via `give_away/3` or
+  by being named the Set's heir (see `new/1`) - the module should include at least one `accept`
+  clause.  For example, if we want a server to inherit Sets after their previous owner dies:
+
+  ```
+  defmodule Receiver do
+    use GenServer
+    alias ETS.Set
+    require ETS.Set
+
+    ...
+
+    Set.accept :owner_crashed, set, _from, state do
+      new_state = Map.update!(state, :crashed_sets, &[set | &1])
+      {:noreply, new_state}
+    end
+  ```
+
+  The first argument is a unique identifier which should match either the "heir_data"
+  in `new/1`, or the "gift" in `give_away/3`.
+  The other arguments declare the variables which may be used in the `do` block:
+  the received Set, the pid of the previous owner, and the current state of the process.
+
+  The return value should be in the form {:noreply, new_state}, or one of the similar
+  returns expected by `handle_info`/`handle_cast`.
+  """
+  defmacro accept(id, table, from, state, do: contents) do
+    quote do
+      require Base
+
+      Base.accept unquote(id), unquote(table), unquote(from), unquote(state) do
+        var!(unquote(table)) = Set.wrap_existing!(unquote(table))
+        unquote(contents)
+      end
+    end
+  end
 end