+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next
+
+* Add match_object/{1,2,3} and match_delete/2
+* Add give_away/3, accept/1, and accept/6 - Thanks @APB9785
+
 ## 0.8.1
 
 * Add Set.fetch/2 - Thanks @christhekeele